requirements/prod.txt
django-guardian==1.3.1
django==1.8.4
djangorestframework-jwt==1.7.2
djangorestframework==3.2.3
psycopg2==2.6.1
